Cover inline image migration gaps
